move type definition to dunder init
