+++ dcspy/aircraft.py
@@ -11,34 +11,8 @@ from typing import Dict, Union, Iterator, Sequence, List
 
 from PIL import Image, ImageDraw, ImageFont
 
-from dcspy import LcdInfo, LcdButton, LcdType, SUPPORTED_CRAFTS, DED_FONT, config
+from dcspy import LcdInfo, LcdButton, LcdType, SUPPORTED_CRAFTS, DED_FONT, config, BiosValue
 from dcspy.sdk import lcd_sdk
-try:
-    from typing import NotRequired
-except ImportError:
-    from typing_extensions import NotRequired
-try:
-    from typing import TypedDict
-except ImportError:
-    from typing_extensions import TypedDict
-
-
-class IntBuffArgs(TypedDict):
-    address: int
-    mask: int
-    shift_by: int
-
-
-class StrBuffArgs(TypedDict):
-    address: int
-    max_length: int
-
-
-class BiosValue(TypedDict):
-    klass: str
-    args: Union[StrBuffArgs, IntBuffArgs]
-    value: Union[int, str]
-    max_value: NotRequired[int]
 
 
 LOG = getLogger(__name__)
